requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
apscheduler
black